refactor(glm52): split the 2.3k-line scheduler.rs into focused modules
Pure code motion along the seams the module doc already names: slot.rs
(per-request state machine), admission.rs (validate/intake/placement),
plan.rs (step shapes, launch-ahead flags, sampling rows), mod.rs (the
coordinator shell), contract_tests.rs (the offline kvbm-contract drive),
testkit.rs (shared fixtures). Scheduler-internal items tighten from
pub(crate) to pub(super) — the crate surface is run_dp8_coordinator only.
The coordinator's two private-field pokes into Glm52SlotState become
mid_prefill()/remaining_prompt() accessors; behavior is unchanged and
every file is now under 750 lines.
